Problem: Postgres extensions need to be reliable
Postgres extensions should not crash the database and also exhibit
correct functional properties when they alter the behaviour of the
database in a significant way.

Doing this is difficult and expensive, especially in C. You can solve a
number of problems by using Rust and [pgrx](https://github.com/pgcentralfoundation/pgrx/)
but it'll be mostly around memory and null safety.

Solution: use Ada to develop extensions

With Ada and SPARK, one can benefit from type specialization, robust
invariants and compile-time correctness verification.

This initial commit brings a prototype of an extension framework with a
simple example. It's very rough on edges, meant only as a proof of
concept.