Problem: Safe_FFI naming
It implies it is safe (duh!); but that's not really accurate. What it
does right now is it allows to exit correctly by winding all the
controlled type values.

Solution: rename it to Checked_FFI

Still not perfect, but it's a bit better, I think.

While at it, handle absense of `gnatpp` better in the codegen tool.