Problem: postgres errors propagate all the way out
This is not how it is supposed to be; if we want to handle Postgres
error inside of our Ada code, we should be able to.

Solution: raise Postgres_Error when Postgres error is occurred

If propagated all the way to the top (Native_Function), then it'll be
converted back to Postgres error.

This change also fixes a number of things that were wrong in setting up
the environment for checked calls.

Codegen is still extremely brittle and for that reason I've used
fully-qualified type names in arguments, to avoid the complexity of
generating package body that has everything referenced correctly.