Problem: using Nullable_Datum is a bit inconvenient
Requires conversion and the one we provided doesn't actually check for
null.

Solution: make it a public record with a discriminant

This forces us to make Any_Datum public for sizing reasons. On the flip
side, making Datum_Int* public makes it easier to do arithmetics on
them!

Also fixed an omission in Ada exception handling.